old/requirements.txt
flask
gunicorn
spotipy
requests
beautifulsoup4
python-dotenv
lxml
curl_cffi
